WidgetFormのSoundCloud周りを修正
Fixes #1141
